Some changes on Front and Backend both

+++ Backend/CarWash-Customer-Service/src/main/java/com/capgemini/carwash/entity/CarDetails.java
@@ -13,7 +13,7 @@ import lombok.Data;
 public class CarDetails {
 
 	@Id
-    private Long car_id;
+    private Integer car_id;
 
     @NaturalId
     @NotEmpty
